+++ graphistry/PlotterBase.py
@@ -258,6 +258,10 @@ class PlotterBase(Plottable):
         # NOTE: See plotter initialization for session bindings & concurrency notes.
         super().__init__(*args, **kwargs)
 
+        self._cypher_entity_projection_kinds = None
+        self._cypher_entity_projection_presence = {}
+        self._cypher_entity_projection_meta = {}
+
         # Bindings
         self._edges : Any = None
         self._nodes : Any = None